pyproject.toml
[project]
name = "hiperhealth"
version = "0.4.0"  # semantic-release
description = "Provide a pipeline for medical encounter empowered by AI"
readme = "README.md"
authors = [
  {name = "Ivan Ogasawara", email = "ivan.ogasawara@gmail.com"}
]
license = { text = "BSD 3 Clause" }
requires-python = ">=3.10,<4"
dependencies = [
    "fhir.resources >=8.1.0",
    "pypdf >=5.4.0",
    "sqlalchemy >=2.0.41",
    "atpublic >=6",
    "questionary >=2.1.0",
    "pillow >=11.3.0",
    "pytesseract >=0.3.13",
    "python-magic >=0.4.27",
    "presidio-analyzer >=2.2.359,<3.0.0",
    "presidio-anonymizer >=2.2.359,<3.0.0",
    "torch >=2",
    "langchain >=0.3.7,<1",
    "langchain-community >=0.3.7,<1",
    "litellm >=1,<2",
    "pyyaml >=6",
    "types-PyYAML >=6",
    # NumPy: macOS stays on 1.x for Python <3.13; others use 2.x
    "numpy >=1.26,<2; sys_platform == 'darwin' and python_version < '3.13'",
    "numpy >=2; sys_platform == 'darwin' and python_version >= '3.13'",
    "numpy >=2,<3; sys_platform != 'darwin'",
]

[project.urls]
Homepage = "https://github.com/hiperhealth/hiperhealth"
Issues = "https://github.com/hiperhealth/hiperhealth/issues"

[project.entry-points."hiperhealth.skills"]
diagnostics = "hiperhealth.skills.diagnostics:DiagnosticsSkill"
privacy = "hiperhealth.skills.privacy:PrivacySkill"
extraction = "hiperhealth.skills.extraction:ExtractionSkill"

[build-system]
requires = ["setuptools>=70.1.0,<82", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
hiperhealth = ["py.typed", "skills/*/hiperhealth.yaml"]

[project.optional-dependencies]
dev = [
  "build >=1.3.0",
  "twine >=6",
  "douki >=0.12.0",
  "pytest >=7.3.2",
  "pytest-cov >=4.1.0",
  "coverage >=7.2.7",
  "pre-commit >=3.3.2",
  "ruff >=0.2.0",
  "mypy >=1.5.0",
  "bandit >=1.7.5",
  "vulture >=2.7",
  "ipython >=8",
  "ipykernel >=6.0.0",
  "Jinja2 >=3.1.2",
  "mkdocs >=1.3",
  "mkdocs-exclude >=1.0.2",
  "mkdocs-jupyter >=0.24.7",
  "mkdocs-literate-nav >=0.4.1",
  "mkdocs-macros-plugin >=0.6.3",
  "mkdocs-material >=8",
  "mkdocstrings >=0.21.2",
  "mkdocstrings-python >= 1.1.2",
  "mkdocs-gen-files >=0.5.0",
  "makim ==1.27.0",
  "python-multipart >=0.0.20",
]

[tool.bandit]
exclude_dirs = ["tests"]
targets = "./"

[tool.mypy]
python_version = "3.10"
allow_subclassing_any = true
check_untyped_defs = true
strict = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
exclude = ["scripts/"]
plugins = [
  "pydantic.mypy",
]

[[tool.mypy.overrides]]
module = "hiperhealth.schema.fhirx"
disable_error_code = [
  "name-defined",
  "valid-type",
]

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
pythonpath = [
    "src",
]
filterwarnings = [
    "error::RuntimeWarning",
    "error::typeguard.TypeHintWarning",
]

[tool.ruff]
target-version = "py310"
line-length = 79
force-exclude = true
src = ["./"]
exclude = [
  'docs',
]
fix = true

[tool.ruff.lint]
select = [
  "E",   # pycodestyle
  "F",   # pyflakes
  "YTT", # flake8-2020
  "RUF", # Ruff-specific rules
  "I001",  # isort
]
extend-select = [
  "UP006",  # Prefer builtin generics in annotations
  "UP007",  # Prefer | syntax for unions and optionals
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.isort]
# Use a single line between direct and from import
lines-between-types = 1

[tool.ruff.format]
quote-style = "single"

[tool.vulture]
exclude = ["tests"]
ignore_decorators = []
ignore_names = []
make_whitelist = true
min_confidence = 80
paths = ["./", "vulture_whitelist.py"]
sort_by_size = true
verbose = false
